# tracer: pre-size payload buffer to optimize memory allocations