--- a/ddtrace/tracer/payload.go
+++ b/ddtrace/tracer/payload.go
@@ -72,6 +72,7 @@ func newPayload() *payload {
 
 // push pushes a new item into the stream.
 func (p *payload) push(t spanList) error {
+	p.buf.Grow(t.Msgsize())
 	if err := msgp.Encode(&p.buf, t); err != nil {
 		return err
 	}