@ajgajg1134 2023-10-31
I took a look at this locally and it appears this is likely because the benchmark for partial flushing is using a somewhat small number (500) to trigger a flush, so the new line here where we expand the payload buffer is causing slightly more memory pressure than before. However compared to with partial flushing OFF there are still significant memory benefits (ran this locally to confirm), given that there's still an improvement and partial flushing is pretty niche I would say this is safe to merge!

(And if we ran the benchmark with a larger partial flush span number like 1000 or 2000 this memory pressure would be even further reduced)